+++ src/commands/PermissionCheckCommand.ts
@@ -18,5 +18,5 @@ import { Mjolnir } from "../Mjolnir";
 
 // !mjolnir verify
 export async function execPermissionCheckCommand(roomId: string, event: any, mjolnir: Mjolnir) {
-    return mjolnir.verifyPermissions(true, true);
+    return mjolnir.protectedRoomsTracker.verifyPermissions(true, true);
 }
