+++ test/integration/detectFederationLagTest.ts
@@ -24,8 +24,8 @@ describe("Test: DetectFederationLag protection", function() {
     beforeEach(async function() {
         // Setup an instance of DetectFederationLag
         this.detector = new DetectFederationLag();
-        await this.mjolnir.registerProtection(this.detector);
-        await this.mjolnir.enableProtection("DetectFederationLag");
+        await this.mjolnir.protectionManager.registerProtection(this.detector);
+        await this.mjolnir.protectionManager.enableProtection("DetectFederationLag");
 
         // Setup a moderator.
         this.moderator = await newTestUser(this.config.homeserverUrl, { name: { contains: "moderator" } });
